refactor(workflow): reorganize step card components
