+++ components/workflow/step-card/details.tsx
